+++ lang/en/messages.php
@@ -5,4 +5,5 @@ declare(strict_types=1);
 return [
     'placeholder_heading' => 'Stencil',
     'placeholder' => 'Stencil placeholder translation.',
+    'dialog_close' => 'Close',
 ];