+++ resources/views/components/dialog/action.blade.php
@@ -1,0 +1,7 @@
+@props([
+    'variant' => 'primary',
+])
+
+<x-stencil::button :variant="$variant" {{ $attributes }}>
+    {{ $slot }}
+</x-stencil::button>