+++ resources/views/components/dialog/cancel.blade.php
@@ -1,0 +1,7 @@
+@props([
+    'variant' => 'outline',
+])
+
+<x-stencil::button :variant="$variant" {{ $attributes->merge(['data-dialog-close' => true]) }}>
+    {{ $slot }}
+</x-stencil::button>