+++ resources/views/components/dialog/close.blade.php
@@ -1,0 +1,8 @@
+<button
+    type="button"
+    {{ $attributes->class(['dialog__close', 'inline-flex'])->merge([
+        'data-dialog-close' => true,
+    ]) }}
+>
+    {{ $slot }}
+</button>