+++ resources/views/components/dialog/footer.blade.php
@@ -1,0 +1,3 @@
+<div {{ $attributes->class(['dialog__footer', 'mt-6 flex flex-col-reverse gap-2 sm:flex-row sm:justify-end']) }} data-dialog-footer>
+    {{ $slot }}
+</div>