+++ resources/views/components/dialog/header.blade.php
@@ -1,0 +1,3 @@
+<div {{ $attributes->class(['dialog__header', 'flex flex-col gap-2 pr-8 text-left']) }} data-dialog-header>
+    {{ $slot }}
+</div>