+++ resources/views/components/dialog/index.blade.php
@@ -1,0 +1,13 @@
+@props([
+    'name' => null,
+])
+
+<div
+    {{ $attributes->class(['dialog', 'contents']) }}
+    data-dialog
+    @if (filled($name))
+        data-dialog-name="{{ $name }}"
+    @endif
+>
+    {{ $slot }}
+</div>